% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as_template}
\alias{as_template}
\title{Modify an input OBIS dataset so that input fields match template types}
\usage{
as_template(x, template = species_template(n = 1, eventDate_type = "date"))
}
\arguments{
\item{x}{tibble of raw GBIF data}

\item{template}{template as a tibble}
}
\value{
tibble with defined columns
}
\description{
Modify an input OBIS dataset so that input fields match template types
}
\seealso{
\href{https://ipt.gbif.org/manual/en/ipt/2.5/occurrence-data}{Darwin Core}
}
