% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_species}
\alias{fetch_species}
\title{Fetch a species from OBIS}
\usage{
fetch_species(
  scientificname = "Alexandrium affine",
  save_file = file_name(scientificname),
  template = species_template(),
  ...
)
}
\arguments{
\item{scientificname}{character, the Latin name for a species}

\item{save_file}{NA or a path specification to save the file}

\item{template}{character or NULL, set to a charcater vector of desired fields, or 
NULL to retrieve all available fields.  Programmatically, it is better to specify the 
desired fields so that for each species the output has the same species.}

\item{...}{other arguments for \code{\link[robis]{occurrence}}}
}
\value{
tibble, possibly empty if a species is not found
}
\description{
Fetch a species from OBIS
}
