% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_path}
\alias{get_path}
\title{Retrieve a path}
\usage{
get_path(..., root = habobis_root())
}
\arguments{
\item{...}{character, file path segments}

\item{root}{the root path}
}
\value{
the full file specification
}
\description{
Retrieve a path
}
