% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habobis-package.R
\docType{package}
\name{habobis-package}
\alias{habobis-package}
\alias{habobis}
\title{habobis package}
\description{
Tools for working with Copernicus data.
}
\details{
Use this package to fetch and manage OBIS data about HAB species. Learn more about 
OBIS [here](https://obis.org/). 

Data is stored locally, and the path maybe semi-hardwired by populating a
text file, `.habobis` in your home directory. The file should have just one line,
}
