% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_species}
\alias{list_species}
\title{List species in database of files}
\usage{
list_species(
  path = get_path(),
  form = c("filename", "scientificname")[1],
  pattern = "^.*\\\\.csv\\\\.gz$"
)
}
\arguments{
\item{path}{chr, path of species data folder}

\item{form}{character, one of 'filename' (default) or 'scientificname'}

\item{pattern}{charcater, regex pattern for \{code{\link[base]{list.files}}}
}
\value{
character vector of species within folder as filename or scientific name
}
\description{
List species in database of files
}
