% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_species}
\alias{read_species}
\title{Read one or more species files}
\usage{
read_species(
  scientificname = "Alexandrium affine",
  fetch = TRUE,
  bind = TRUE,
  template = species_template(n = 1, eventDate_type = "date"),
  fill_names = TRUE,
  form = c("tibble", "sf")[1],
  date_required = TRUE
)
}
\arguments{
\item{scientificname}{character, one or more species to read. If this contains "all" then
all of the species available are read.}

\item{fetch}{logical, if a local instance of the species is not found, should
it be fetched from obis?}

\item{bind}{logical, if TRUE and multiple species are read, bind them into one data frame}

\item{template, }{NULL or a tibble, if a tibble then trim down to the bare 
essentials defined by the template}

\item{fill_names}{logical, if TRUE fill any missing scientificname values automatically}

\item{form}{character, one of 'tibble' or 'sf' to modify the output}

\item{date_required}{logical, if TRUE all eventDate values must be non-NA}
}
\value{
list of tibbles, or if \code{bind} is TRUE a single tibble
}
\description{
Read one or more species files
}
