% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{species_template}
\alias{species_template}
\title{Generate a dummy template of data}
\usage{
species_template(n = 1)
}
\arguments{
\item{n}{numeric, the number of rows to create}
}
\value{
tibble
}
\description{
Generate a dummy template of data
}
